% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelRegression-package.R
\docType{package}
\name{ParallelRegression-package}
\alias{ParallelRegression}
\alias{ParallelRegression-package}
\title{ParallelRegression: What the package does (short line)}
\description{
Perform logistic regression in parallel and provide functionality to compare approaches.
}
\keyword{internal}
