\name{ParallelRegression-package}
\alias{ParallelRegression-package}
\alias{ParallelRegression}
\docType{package}
\title{
\packageTitle{ParallelRegression}
}
\description{
\packageDescription{ParallelRegression}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ParallelRegression}
\packageIndices{ParallelRegression}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ParallelRegression}

Maintainer: \packageMaintainer{ParallelRegression}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
