% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ParLR}
\alias{ParLR}
\title{Parallel Logistic Regresssion}
\usage{
ParLR(x, y, ncores = 1L, comm = 0L)
}
\arguments{
\item{x}{the X matrix in logistic regression}

\item{y}{the response matrix in logistic regression}

\item{ncores}{the number of cores to use}
}
\description{
Parallel Logistic Regresssion
}
